fix metric
